+++ src/main/java/owl/automaton/EdgeTreeAutomatonMixin.java
@@ -23,61 +23,41 @@ import static owl.automaton.Automaton.PreferredEdgeAccess.EDGES;
 import static owl.automaton.Automaton.PreferredEdgeAccess.EDGE_MAP;
 import static owl.automaton.Automaton.PreferredEdgeAccess.EDGE_TREE;
 
-import de.tum.in.naturals.bitset.BitSets;
 import java.util.BitSet;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import owl.automaton.acceptance.OmegaAcceptance;
 import owl.automaton.edge.Edge;
 import owl.collections.ValuationSet;
-import owl.collections.ValuationTree;
-import owl.factories.ValuationSetFactory;
 
 /**
- * Mixin interface for implementing an automaton by {@link Automaton#edges(Object, BitSet)}.
+ * Mixin interface for implementing an automaton by {@link Automaton#edgeTree(Object)}.
  *
- * <p>It is impossible to implement the incompatible sister interface
- * {@link EdgeMapAutomatonMixin} and the compiler will reject the code, since there are
- * conflicting defaults for {@link Automaton#preferredEdgeAccess()}.</p>
+ * <p>It is impossible to implement the incompatible sister interface {@link EdgesAutomatonMixin}
+ * and the compiler will reject the code, since there are conflicting defaults for
+ * {@link Automaton#preferredEdgeAccess()}.</p>
  *
  * @param <S> the type of the states of the automaton
  * @param <A> the type of the omega-acceptance condition of the automaton
  */
-public interface EdgesAutomatonMixin<S, A extends OmegaAcceptance> extends Automaton<S, A> {
+public interface EdgeTreeAutomatonMixin<S, A extends OmegaAcceptance> extends Automaton<S, A> {
 
-  List<PreferredEdgeAccess> ACCESS_MODES = List.of(EDGES, EDGE_MAP, EDGE_TREE);
-
-  @Override
-  default Set<Edge<S>> edges(S state) {
-    Set<Edge<S>> edges = new HashSet<>();
-
-    for (BitSet valuation : BitSets.powerSet(factory().alphabetSize())) {
-      edges.addAll(edges(state, valuation));
-    }
-
-    return edges;
-  }
+  List<PreferredEdgeAccess> ACCESS_MODES = List.of(EDGE_TREE, EDGE_MAP, EDGES);
 
   @Override
   default Map<Edge<S>, ValuationSet> edgeMap(S state) {
-    ValuationSetFactory factory = factory();
-    Map<Edge<S>, ValuationSet> labelledEdges = new HashMap<>();
-
-    for (BitSet valuation : BitSets.powerSet(factory.alphabetSize())) {
-      for (Edge<S> edge : edges(state, valuation)) {
-        labelledEdges.merge(edge, factory.of(valuation), ValuationSet::union);
-      }
-    }
-
-    return labelledEdges;
+    return edgeTree(state).inverse(factory());
   }
 
   @Override
-  default ValuationTree<Edge<S>> edgeTree(S state) {
-    return factory().inverse(edgeMap(state));
+  default Set<Edge<S>> edges(S state) {
+    return edgeTree(state).values();
+  }
+
+  @Override
+  default Set<Edge<S>> edges(S state, BitSet valuation) {
+    return edgeTree(state).get(valuation);
   }
 
   @Override